pyproject.toml
[project]
name = "AIA"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.13.0",
    "funasr>=1.3.1 ; sys_platform == 'linux'",
    "langchain-chroma>=1.1.0 ; sys_platform == 'linux'",
    "langchain-community>=0.4.1 ; sys_platform == 'linux'",
    "langchain-huggingface==1.0.0 ; sys_platform == 'linux'",
    "lark-oapi>=1.5.3",
    "loguru>=0.7.3",
    "modelscope>=1.35.3",
    "openai>=2.31.0",
    "pathvalidate>=3.3.1",
    "pyyaml>=6.0.3",
    "readability-lxml>=0.8.4.1",
    "sentence-transformers>=5.4.0 ; sys_platform == 'linux'",
    "tiktoken>=0.12.0",
    "torchaudio==2.10.0 ; sys_platform == 'linux'",
]


[[tool.uv.index]]
name = "tsinghua"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple"
default = true


[dependency-groups]
dev = [
    "ipython>=9.12.0",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]

